clean up unit tests (#93)
* use python 3.8
no need for conftest at the moment
relax test a bit

* still need outside conftest

* supress sep np.int warning until resolved in sep

* configs will be revived later

* no need to keep around broken notebooks (we can download them later or look at them in github)

* test reorganized so more understandable now

* black

* supress SHA warning

* slight error
