+++ .github/workflows/config.yml
@@ -14,13 +14,14 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.7]
+        python-version: [3.8]
 
     steps:
       - name: Checkout github repo (+ download lfs dependencies)
         uses: actions/checkout@v2
         with:
           lfs: true
+          fetch-depth: 2
 
       - name: Checkout LFS objects
         run: git lfs checkout
