+++ .gitignore
@@ -1,6 +1,5 @@
 *.ipynb_checkpoints
 *__pycache__
-*.idea
 /my-notebooks/
 .vscode/
 .pytest_cache/
