+++ pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
 filterwarnings =
-    ignore:.*the imp module is deprecated.*:DeprecationWarning
+    ignore:the imp module is deprecated:DeprecationWarning
+    ignore:`np.int` is a deprecated alias:DeprecationWarning
