--- tests/test-config.yaml
@@ -1,96 +1,0 @@
-# This YAML file is to be used with btk_input to simulate objects with
-# BlendingTooKit and perform detection/deblending/measurement.
-
-#This config file is to be used in the btk unit tests.
-
-# Modify the following values; specify directory locations and name of your function.
-# If value are set to None, then the default btk parameters are used.
-
-user_input:
-    # Location to load dataset from
-    data_dir: data/
-    # Location where btk test output will be saved
-    output_dir: tests
-    # Enter file name containing user function to perform detection/deblending/measurement
-    utils_filename: None
-    output_name: unit_test_output # btk output will be saved in a directory with this name inside output_dir
-    utils_input:
-        measure_function: None
-        metrics_function: None
-
-
-# Below are the parameter values specifying 3 kinds of simulations.
----
-simulation: two_gal
-config:
-    catalog: 'sample_input_catalog.fits'
-    max_number: 2
-    batch_size: 8
-    stamp_size: 24
-    survey_name: "LSST"
-    add_noise: True
-    bands:
-        - 'u'
-        - 'g'
-        - 'r'
-        - 'i'
-        - 'z'
-        - 'y'
-    min_snr: 0.05
-    seed: 0
-    add: None
-    selection_function: None
-    sampling_function: None
-    obs_conditions: None
-    test_size: 2
-
----
-simulation: multi_gal
-config:
-    catalog: 'sample_input_catalog.fits'
-    max_number: 10
-    batch_size: 8
-    stamp_size: 24
-    survey_name: "LSST"
-    add_noise: True
-    bands:
-        - 'u'
-        - 'g'
-        - 'r'
-        - 'i'
-        - 'z'
-        - 'y'
-    min_snr: 0.05
-    seed: 1
-    add: None
-    selection_function: basic_selection_function
-    sampling_function: DefaultSampling
-    obs_conditions: None
-    test_size: 2
-
----
-
-simulation: group
-config:
-    catalog: 'sample_group_input_catalog.fits'
-    wld_catalog:
-    max_number: 10
-    batch_size: 8
-    stamp_size: 24
-    survey_name: "LSST"
-    add_noise: True
-    bands:
-        - 'u'
-        - 'g'
-        - 'r'
-        - 'i'
-        - 'z'
-        - 'y'
-    min_snr: 0.05
-    seed: 1
-    add:
-        wld_catalog_name: 'data/sample_group_catalog.fits'
-    selection_function: None
-    sampling_function: GroupSamplingFunction
-    obs_conditions: None
-    test_size: 2
