+++ tests/test_catalog.py
@@ -4,5 +4,5 @@ CATALOG_PATH = "data/sample_input_catalog.fits"
 
 
 def test_verbose():
-    """For coverage"""
+    # for coverage.
     CatsimCatalog.from_file(CATALOG_PATH, verbose=True)
